RATIS-1476. Refactor the snapshot installation code to a new class.
